+++ cli/commands/entity/list.go
@@ -7,6 +7,7 @@ import (
 	"github.com/sensu/sensu-go/cli"
 	"github.com/sensu/sensu-go/cli/commands/flags"
 	"github.com/sensu/sensu-go/cli/commands/helpers"
+	"github.com/sensu/sensu-go/cli/commands/timeutil"
 	"github.com/sensu/sensu-go/cli/elements/table"
 	"github.com/sensu/sensu-go/types"
 	"github.com/spf13/cobra"
@@ -76,7 +77,7 @@ func printToTable(results interface{}, writer io.Writer) {
 			Title: "Last Seen",
 			CellTransformer: func(data interface{}) string {
 				entity, _ := data.(types.Entity)
-				return helpers.HumanTimestamp(entity.LastSeen)
+				return timeutil.HumanTimestamp(entity.LastSeen)
 			},
 		},
 	})