+++ cli/commands/silenced/create.go
@@ -14,18 +14,12 @@ func CreateCommand(cli *cli.SensuCli) *cobra.Command {
 		Use:          "create",
 		Short:        "create a silenced entry",
 		SilenceUsage: true,
+		Args:         cobra.ExactArgs(0),
 		RunE: func(cmd *cobra.Command, args []string) error {
 			flags := cmd.Flags()
 			isInteractive := flags.NFlag() == 0
 
 			opts := newSilencedOpts()
-
-			if len(args) > 0 {
-				opts.Subscription = args[0]
-			}
-			if len(args) > 1 {
-				opts.Check = args[1]
-			}
 
 			opts.Org = cli.Config.Organization()
 			opts.Env = cli.Config.Environment()
@@ -37,6 +31,12 @@ func CreateCommand(cli *cli.SensuCli) *cobra.Command {
 			} else {
 				if err := opts.withFlags(flags); err != nil {
 					return err
+				}
+				if opts.Check == "" && opts.Subscription == "" {
+					return fmt.Errorf("must specify --check or --subscription")
+				}
+				if opts.Reason == "" {
+					return fmt.Errorf("must specify --reason")
 				}
 			}
 			var silenced types.Silenced
@@ -60,8 +60,7 @@ func CreateCommand(cli *cli.SensuCli) *cobra.Command {
 	_ = cmd.Flags().StringP("expire", "e", expireDefault, "expiry in seconds")
 	_ = cmd.Flags().StringP("subscription", "s", "", "silence subscription")
 	_ = cmd.Flags().StringP("check", "c", "", "silence check")
-	_ = cmd.Flags().StringP("begin", "b", beginDefault, "silence begin in epoch time")
-	_ = cmd.MarkFlagRequired("reason")
+	_ = cmd.Flags().StringP("begin", "b", beginDefault, "silence begin in human readable time (Format: Jan 02 2006 3:04PM MST)")
 
 	return cmd
 }