+++ cli/commands/silenced/create_test.go
@@ -46,10 +46,11 @@ func TestCreateCommandRunEClosureWithAllFlags(t *testing.T) {
 	client.On("CreateSilenced", mock.Anything).Return(nil)
 
 	cmd := CreateCommand(cli)
+	require.NoError(t, cmd.Flags().Set("reason", "just because"))
 	require.NoError(t, cmd.Flags().Set("expire", "5"))
 	require.NoError(t, cmd.Flags().Set("expire-on-resolve", "false"))
 	require.NoError(t, cmd.Flags().Set("subscription", "weeklyworldnews"))
-	require.NoError(t, cmd.Flags().Set("begin", "1257894000"))
+	require.NoError(t, cmd.Flags().Set("begin", "Jan 02 2006 3:04PM MST"))
 	out, err := test.RunCmd(cmd, []string{})
 	require.NoError(t, err)
 	assert.Regexp("OK", out)
@@ -63,6 +64,7 @@ func TestCreateCommandRunEClosureWithDeps(t *testing.T) {
 	client.On("CreateSilenced", mock.AnythingOfType("*types.Silenced")).Return(nil)
 
 	cmd := CreateCommand(cli)
+	require.NoError(t, cmd.Flags().Set("reason", "just because"))
 	require.NoError(t, cmd.Flags().Set("expire", "5"))
 	require.NoError(t, cmd.Flags().Set("expire-on-resolve", "false"))
 	require.NoError(t, cmd.Flags().Set("subscription", "weeklyworldnews"))
@@ -80,6 +82,7 @@ func TestCreateCommandRunEClosureWithServerErr(t *testing.T) {
 	client.On("CreateSilenced", mock.AnythingOfType("*types.Silenced")).Return(errors.New("whoops"))
 
 	cmd := CreateCommand(cli)
+	require.NoError(t, cmd.Flags().Set("reason", "just because"))
 	require.NoError(t, cmd.Flags().Set("expire", "5"))
 	require.NoError(t, cmd.Flags().Set("expire-on-resolve", "false"))
 	require.NoError(t, cmd.Flags().Set("subscription", "weeklyworldnews"))
@@ -87,5 +90,19 @@ func TestCreateCommandRunEClosureWithServerErr(t *testing.T) {
 	out, err := test.RunCmd(cmd, []string{})
 	require.Error(t, err)
 	assert.Equal("whoops", err.Error())
+	assert.Empty(out)
+}
+
+func TestCreateCommandRunEClosureWithMissingRequiredFlags(t *testing.T) {
+	assert := assert.New(t)
+
+	cli := test.NewMockCLI()
+	client := cli.Client.(*client.MockClient)
+	client.On("CreateSilenced", mock.AnythingOfType("*types.Silenced")).Return(errors.New("error"))
+
+	cmd := CreateCommand(cli)
+	require.NoError(t, cmd.Flags().Set("expire", "5"))
+	out, err := test.RunCmd(cmd, []string{})
+	require.Error(t, err)
 	assert.Empty(out)
 }