+++ CHANGELOG.md
@@ -7,6 +7,22 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
 ### Added
+- Add .gitattributes file with merge stratety for the Changelog.
+- Context switcher added for dashboard
+- Add api support for adhoc check requests.
+
+### Changed
+- Silenced `begin` supports human readable time (Format: Jan 02 2006 3:04PM MST)
+in `sensuctl` with optional timezone. Stores the field as unix epoch time.
+- Increased the timeout in the store's watchers tests
+- Incremental retry mechanism when waiting for agent and backend in e2e tests.
+
+### Fixed
+- Fixed required flags in CLI `check` and `silenced` creation and added better
+error checking.
+
+## [2.0.0-alpha.16] - 2018-02-07
+### Added
 - Add an e2e test for proxy check requests.
 - Add integration tests to our CI.
 
@@ -20,11 +36,18 @@ finished before its allotted execution timeout.
 - Properly log errors whenever a check request can't be published.
 - Fixed some build tags for tests using etcd stores.
 - Keepalive monitors now get updated with changes to a keepalive timeout.
+- Prevent tests timeout in queue package
+- Prevent tests timeout in ring package
+- Fixed a bug in the queue package where timestamps were not parsed correctly.
+- Fixed Ring's Next method hanging in cases where watch events are not propagated.
 
 ### Changed
 - Queues are now durable.
 - Refactoring of the check scheduling integration tests.
 - CLI resource delete confirmation is now `(y/N)`.
+
+### Removed
+- Dependency github.com/chzyer/readline
 
 ## [2.0.0-alpha.15] - 2018-01-30
 ### Added